requirements.txt
numpy==1.11.1
python-dateutil==2.5.3
matplotlib==1.5.1
pandas==0.18.1